Rename ReplyFeedback to ArticleReplyFeedbackControl This includes display (button groups, other's reasons) and inputs (upvote, downvote, reason)
