+++ components/ArticleReplyFeedbackControl/ArticleReplyFeedbackControl.js
@@ -2,82 +2,17 @@ import React, { useState } from 'react';
 import { t } from 'ttag';
 import { gql, useMutation } from '@apollo/client';
 import { makeStyles, withStyles } from '@material-ui/core/styles';
-import {
-  Tabs,
-  Tab,
-  ButtonGroup,
-  Button,
-  Box,
-  Popover,
-  Typography,
-  SvgIcon,
-} from '@material-ui/core';
+import { Tabs, Tab, Box, Popover, Typography } from '@material-ui/core';
 import Snackbar from '@material-ui/core/Snackbar';
 import CloseIcon from '@material-ui/icons/Close';
 import Avatar from 'components/AppLayout/Widgets/Avatar';
-import useCurrentUser from 'lib/useCurrentUser';
+import { ThumbUpIcon, ThumbDownIcon } from 'components/icons';
+import ButtonGroupDisplay from './ButtonGroupDisplay';
 import cx from 'clsx';
 
 const useStyles = makeStyles(theme => ({
   root: {
     display: 'flex',
-  },
-  vote: {
-    display: 'flex',
-    alignItems: 'center',
-    borderRadius: 45,
-    padding: '0px 8px',
-    marginRight: 3,
-    outline: 'none',
-    cursor: 'pointer',
-    border: `1px solid ${theme.palette.secondary[100]}`,
-    color: theme.palette.secondary[100],
-    background: theme.palette.common.white,
-    [theme.breakpoints.up('md')]: {
-      padding: '0 18px',
-      marginRight: 10,
-    },
-    '&:hover': {
-      border: `1px solid ${theme.palette.secondary[300]}`,
-      color: theme.palette.secondary[300],
-    },
-  },
-  thumbIcon: {
-    fontSize: 15,
-    margin: '0 2px',
-  },
-  buttonGroup: {
-    display: 'flex',
-    alignItems: 'center',
-    whiteSpace: 'nowrap',
-    '& $vote': {
-      marginRight: 0,
-      display: 'inline-flex',
-      color: theme.palette.secondary[200],
-      fontSize: 14,
-      textTransform: 'none',
-      '&:first-child': {
-        borderTopRightRadius: 0,
-        borderBottomRightRadius: 0,
-        borderRight: 0,
-        paddingRight: 0,
-      },
-      '&:last-child': {
-        borderTopLeftRadius: 0,
-        borderBottomLeftRadius: 0,
-      },
-      '&:not(:first-child):not(:last-child)': {
-        borderRadius: 0,
-        borderLeft: 0,
-        borderRight: 0,
-      },
-      '&:disabled': {
-        cursor: 'not-allowed',
-      },
-    },
-  },
-  voted: {
-    color: `${theme.palette.primary[500]} !important`,
   },
   popover: {
     position: 'relative',
@@ -131,17 +66,6 @@ const useStyles = makeStyles(theme => ({
   },
 }));
 
-const ThumbUpIcon = props => (
-  <SvgIcon {...props} viewBox="0 0 17 15">
-    <path d="M1 13.8672H3.65455V6.51612H1V13.8672ZM15.6 7.1287C15.6 6.45486 15.0027 5.90353 14.2727 5.90353H10.0852L10.7156 3.10401L10.7355 2.90798C10.7355 2.65682 10.6227 2.42403 10.4435 2.25863L9.74009 1.61542L5.37336 5.65237C5.12782 5.8729 4.98182 6.17919 4.98182 6.51612V12.642C4.98182 13.3158 5.57909 13.8672 6.30909 13.8672H12.2818C12.8326 13.8672 13.3038 13.5609 13.5029 13.1198L15.5071 8.80107C15.5668 8.66017 15.6 8.51315 15.6 8.35388V7.1287Z" />
-  </SvgIcon>
-);
-
-const ThumbDownIcon = props => (
-  <SvgIcon {...props} viewBox="0 0 17 15">
-    <path d="M10.2909 1.46155H4.31818C3.76736 1.46155 3.29618 1.76784 3.09709 2.2089L1.09291 6.52765C1.03318 6.66854 1 6.81556 1 6.97483V8.20001C1 8.87386 1.59727 9.42518 2.32727 9.42518H6.51482L5.88436 12.2247L5.86445 12.4207C5.86445 12.6719 5.97727 12.9047 6.15646 13.0701L6.85991 13.7133L11.2333 9.67635C11.4722 9.45581 11.6182 9.14952 11.6182 8.8126V2.68672C11.6182 2.01288 11.0209 1.46155 10.2909 1.46155ZM12.9455 1.46155V8.8126H15.6V1.46155H12.9455Z" />
-  </SvgIcon>
-);
 const CustomTab = withStyles({
   root: {
     position: 'relative',
@@ -179,17 +103,18 @@ const Feedback = withStyles(theme => ({
 
 // Subset of fields that needs to be updated after login
 //
-const ArticleReplyFeedbackForUser = gql`
-  fragment ArticleReplyFeedbackForUser on ArticleReply {
+const ArticleReplyFeedbackControlDataForUser = gql`
+  fragment ArticleReplyFeedbackControlDataForUser on ArticleReply {
     articleId
     replyId
     ownVote
+    ...ButtonGroupDisplayArticleReplyForUser
   }
+  ${ButtonGroupDisplay.fragments.ButtonGroupDisplayArticleReplyForUser}
 `;
 
-const ArticleReplyFeedbackData = gql`
-  fragment ArticleReplyFeedbackData on ArticleReply {
-    ...ArticleReplyFeedbackForUser
+const ArticleReplyFeedbackControlData = gql`
+  fragment ArticleReplyFeedbackControlData on ArticleReply {
     positiveFeedbackCount
     negativeFeedbackCount
     ownVote
@@ -203,8 +128,11 @@ const ArticleReplyFeedbackData = gql`
         avatarUrl
       }
     }
+    ...ArticleReplyFeedbackControlDataForUser
+    ...ButtonGroupDisplayArticleReply
   }
-  ${ArticleReplyFeedbackForUser}
+  ${ArticleReplyFeedbackControlDataForUser}
+  ${ButtonGroupDisplay.fragments.ButtonGroupDisplayArticleReply}
 `;
 
 const CREATE_REPLY_FEEDBACK = gql`
@@ -220,31 +148,27 @@ const CREATE_REPLY_FEEDBACK = gql`
       vote: $vote
       comment: $comment
     ) {
-      ...ArticleReplyFeedbackData
+      ...ArticleReplyFeedbackControlData
     }
   }
-  ${ArticleReplyFeedbackData}
+  ${ArticleReplyFeedbackControlData}
 `;
 
-function ReplyFeedback({
-  articleId,
-  replyId,
-  positiveFeedbackCount,
-  negativeFeedbackCount,
-  ownVote,
-  feedbacks = [],
-  reply = {},
-  className,
-}) {
-  const { user, text } = reply;
-
+/**
+ *
+ * @param {ArticleReply} props.articleReply - ArticleReply from API
+ * @param {Reply} props.reply - the reply instance of current articleReply.
+ *   Isolated because not all use case have reply nested under articleReply.
+ * @param {string?} props.className
+ */
+function ArticleReplyFeedbackControl({ articleReply, reply = {}, className }) {
+  const classes = useStyles();
   const [vote, setVote] = useState(null);
   const [reason, setReason] = useState('');
   const [reasonsPopoverAnchorEl, setReasonsPopoverAnchorEl] = useState(null);
   const [votePopoverAnchorEl, setVotePopoverAnchorEl] = useState(null);
   const [tab, setTab] = useState(0);
   const [showReorderSnack, setReorderSnackShow] = useState(false);
-  const currentUser = useCurrentUser();
   const [createReplyFeedback, { loading: updatingReplyFeedback }] = useMutation(
     CREATE_REPLY_FEEDBACK,
     {
@@ -276,38 +200,14 @@ function ReplyFeedback({
     setVote(null);
   };
 
-  const classes = useStyles();
-
-  // Note that currentUser and user may be undefined or null (when appId mismatch)
-  // Both case should not consider as ownArticleReply
-  //
-  const isOwnArticleReply = currentUser && user && currentUser.id === user.id;
-
   return (
     <div className={cx(classes.root, className)}>
-      <ButtonGroup className={classes.buttonGroup} data-ga="Number display">
-        <Button
-          className={cx(classes.vote, ownVote === 'UPVOTE' && classes.voted)}
-          onClick={e => openVotePopover(e, 'UPVOTE')}
-          disabled={isOwnArticleReply}
-          data-ga="Upvote"
-        >
-          {positiveFeedbackCount}
-          <ThumbUpIcon className={classes.thumbIcon} />
-        </Button>
-        <Button
-          className={cx(classes.vote, ownVote === 'DOWNVOTE' && classes.voted)}
-          onClick={e => openVotePopover(e, 'DOWNVOTE')}
-          disabled={isOwnArticleReply}
-          data-ga="Downvote"
-        >
-          {negativeFeedbackCount}
-          <ThumbDownIcon className={classes.thumbIcon} />
-        </Button>
-        <Button className={classes.vote} onClick={openReasonsPopover}>
-          {t`See Reasons`}
-        </Button>
-      </ButtonGroup>
+      <ButtonGroupDisplay
+        articleReply={articleReply}
+        onVoteUp={e => openVotePopover(e, 'UPVOTE')}
+        onVoteDown={e => openVotePopover(e, 'DOWNVOTE')}
+        onReasonClick={openReasonsPopover}
+      />
       <Popover
         id={`reply-reasons-${reply.id}`}
         open={!!reasonsPopoverAnchorEl}
@@ -326,7 +226,7 @@ function ReplyFeedback({
         >
           <CloseIcon />
         </button>
-        {text}
+        {reply.text}
         <Tabs
           value={tab}
           onChange={(e, value) => setTab(value)}
@@ -336,11 +236,11 @@ function ReplyFeedback({
         >
           <CustomTab
             icon={<ThumbUpIcon />}
-            label={t`Helpful ${positiveFeedbackCount}`}
+            label={t`Helpful ${articleReply.positiveFeedbackCount}`}
           />
           <CustomTab
             icon={<ThumbDownIcon />}
-            label={t`Not Helpful ${negativeFeedbackCount}`}
+            label={t`Not Helpful ${articleReply.negativeFeedbackCount}`}
           />
         </Tabs>
         <Box
@@ -415,11 +315,9 @@ function ReplyFeedback({
   );
 }
 
-ReplyFeedback.displayName = 'ReplyFeedback';
-
-ReplyFeedback.fragments = {
-  ArticleReplyFeedbackData,
-  ArticleReplyFeedbackForUser,
+ArticleReplyFeedbackControl.fragments = {
+  ArticleReplyFeedbackControlData,
+  ArticleReplyFeedbackControlDataForUser,
 };
 
-export default ReplyFeedback;
+export default ArticleReplyFeedbackControl;
