+++ components/ArticleReplyFeedbackControl/index.js
@@ -1,0 +1,2 @@
+import ArticleReplyFeedbackControl from './ArticleReplyFeedbackControl';
+export default ArticleReplyFeedbackControl;
